Status - Unable to display status output on Windows source and binary build #4339

+++ hummingbot/core/data_type/limit_order.pxd
@@ -7,7 +7,7 @@ cdef class LimitOrder:
     cdef:
         CPPLimitOrder _cpp_limit_order
     cdef long c_age(self)
-    cdef long c_age_til(self, long start_timestamp)
+    cdef long c_age_til(self, long long start_timestamp)
 
 
 cdef LimitOrder c_create_limit_order_from_cpp_limit_order(const CPPLimitOrder cpp_limit_order)
